python3.X - build tkinter with python rather than separately.
We were building py3-tkinter separate from the python build.
The big issue with this was that we were only providing
one version of it (only 3.11, since that was being manually updated).

Have to add explicit tcl and tk dependencies because melange does
not seem to recognize the dependency that
_tkinter.cpython-312-x86_64-linux-gnu.so (310, 311, 312, 313) has.

```
$ cd /usr/lib/python3.12
        linux-vdso.so.1 (0x00007ffdc66f6000)
        libtk8.6.so => not found
        libtcl8.6.so => not found
        libc.so.6 => /lib/libc.so.6 (0x000078bac419b000)
        /lib64/ld-linux-x86-64.so.2 (0x000078bac43a3000)
```