fix(chromium): Continue using release monitor (#29215)
As we switched to the git poller, we began picking up updates for the
dev releases of Chromium instead of only the stable releases. Updates
jumped from 129 to 131 (the current dev channel) which we should not
track

Revert to using release monitor for now. Ideally, we'd want to somehow
track the current major release - 2 with the poller in the future

The current stable release of Chromium can be found here:


https://versionhistory.googleapis.com/v1/chrome/platforms/linux/channels/stable/versions/all/releases

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>