ruby3.2-systemd-journal - bump to version 2.0.0
There is currently not a tag in the upstream git repo for version 2.0.0
https://github.com/ledbettj/systemd-journal/issues/101

But 2.0.0 is required for fluent-plugin-systemd v1.1.0.

So for the time being, checkout a branch.  This cause fail to build
from source (FTBFS) at the point when there is a new commit on master.

Hopefully at that point there will be an upstream tag.