datawire-envoy: add setcap subpackage (#29572)
This adds a subpackage that runs setcap to enable certain capabilities
on the binary. This is a requirement for downstream Ambassador
dependencies (see

https://github.com/emissary-ingress/emissary/blob/ac2dc64c6621cd8ec5617f3328544364bdd3fb01/build-aux/Dockerfile#L129-L132).

<!---
Provide a short summary in the Title above. Examples of good PR titles:
* "ruby-3.1: new package"
* "haproxy: fix CVE-2014-123456"
-->

<!--
Please include references to any related issues or delete this section
otherwise.
 -->

Related: https://github.com/chainguard-dev/image-requests/issues/3494