updated pkgconf test pipeline to be more portable (#29584)
This pkgconf test pipeline does a better job of checking what should be
working (and skipping things that should not).

There are two sample tests added... One is to openssl.yaml which
succeeds. While the other is to libarchive.yaml fails. However, the
libarchive failure actually looks like a real failure that needs to be
addressed.

See:
```

make test/openssl
make test/libarchive
```