python/multi-version batch 08, Add py/one-python pipeline
Convert 9 packages to multi-version python.

This 'one-python' pipeline will create a 'python3' in the PATH pointing
to the one /usr/bin/python3.XX that is present.  That way the caller can
be blissfully ignorant that their package didn't actually need 'python3'
installed at all.

This makes an improvement here on
https://github.com/wolfi-dev/os/issues/26818