New package libsdl3
Whilst it is pre-release, it has declared to have stable abi https://github.com/libsdl-org/SDL/releases/tag/preview-3.1.3

And we are starting to see cutting edge software releases that require
libsdl3. It is likely that libsdl2 will be required for a long time,
even when libsdl3 goes stable.

Package both during this transition period.

Needed for:
- https://github.com/wolfi-dev/os/pull/27343