druid avro fix for GHSA-r7pg-v2c8-mfg3 (#30615)
Resolve what can be fixed for druid/GHSA-r7pg-v2c8-mfg3 there still is a
avro dependency as a part of hadoop-client-runtime however we are not
able to remedy that. The avro dependency is referred throughout the
various druid build components so it is set as a property variable in
the parent pom.xml. To bump that env variable we use the maven/pombump
properties flag.