apache-nifi/GHSA-r7pg-v2c8-mfg3 fix (#30620)
Remediating apache-nifi GHSA-r7pg-v2c8-mfg3 regarding avro 1.11.3 is
fixed by bumping the version to 1.11.4. In this instance the property
avro.version is used in more than one place in the parent pom.xml file
so it is set as a property. The following changes are the current
implementation of bumping a maven project's property value.

---------

Signed-off-by: jamie-albert <jamie.albert@chainguard.dev>