ingress-nginx-opentelemetry: add nodeps subpackage
This is the same as the main package, but without any dependencies.
This is somewhat counterintuitive, since this results in a package that is unusable by itself.

ingress-nginx ships this in an image but doesn't expect it to work until it's copied over
into another image that already has already has all of the other dependencies installed.
Including the other dependencies is unnecessary, since this image exists only to move bits around,
not to run any code.