make test-debug target more isolated
previously via https://github.com/wolfi-dev/os/pull/30563 we introduced
test-debug
make target which was calling make test/foo with one MELANGE_EXTRA_OPTS
which is
--interactive. This was working good but there are some usecases where
we also need
to pass other MELANGE_EXTRA_OPTS. For example, one that seems to
commonly used is
--runner=docker to change the runner.
With the previous patch, given we were hardcoding MELANGE_EXTRA_OPTS to
--interactive
only, it was not possible to pass --runner=docker somehow.

this commit isolates the test-debug target and uses
MELANGE_DEBUG_TEST_OPTS to pass in
--interactive while running the tests. This way we can keep using the
existing
MELANGE_EXTRA_OPTS to override the runner.

Signed-off-by: kranurag7 <81210977+kranurag7@users.noreply.github.com>