ant: remove runtime dependency on jre
Note that all other java tools do not depend on any specific jre, and
can run with any.

If we really want build tools to pull in any jre, we should update
runtime dependency on all of them to depend on default-jre virtual
package which can be satisfied by any jre.

Currently because of this runtime dependency ant images for jdk 11+
pull in duplicate jre's.