test(prism): test the mock and proxy servers
Adding more functional tests to the prism package. For the mock server, we're testing the mocked endpoints and the request validation feature. For the proxy server, we're testing request forwarding, and its ability to mock requests that are not implemented upstream. Also changed the test spec files to make them more focused.

https://github.com/chainguard-dev/image-requests/issues/3751
Signed-off-by: Sam Singh Anantha <samsinghanantha@gmail.com>