ingress-nginx-opentelemetry: add nodeps subpackage (#30669)
This is the same as the main package, but without any dependencies. This
is somewhat counterintuitive, since this results in a package that is
unusable by itself.

ingress-nginx ships this in an image but doesn't expect it to work until
it's copied over into another image that already has already has all of
the other dependencies installed. Including the other dependencies is
unnecessary, since this image exists only to move bits around, not to
run any code.

<!---
Provide a short summary in the Title above. Examples of good PR titles:
* "ruby-3.1: new package"
* "haproxy: fix CVE-2014-123456"
-->


Related: https://github.com/chainguard-dev/image-requests/issues/4407,
https://github.com/chainguard-images/images-private/pull/5550