py3-keras multi-version (#30881)
Note that there is no test added here.
Import test fails.  Per pypi:

> To use keras, you should also install the backend of choice:
> tensorflow, jax, or torch. Note that tensorflow is required for using
> certain Keras 3 features: certain preprocessing layers as well as
> tf.data pipelines.

The idea is that the user would supply the correct backend and then can
configure that via KERAS_BACKEND or config file.

tensorflow is not built multi-versioned yet.
wolfi does not have torch or jax builds.

So we're not really able to even import-test this at the moment.

Also note, while that seems terrible, this is strictly better than what
_was_ there, which had no test and lack of _identified_ runtime
dependencies.