Updated commons-io dependency in required opensearch-2 subpackages: GHSA-78wr-2p64-hpwj (#30995)
This is a more complicated package build than what is regularly seen,
opensearch-2 has many plugins that exist in both its own repository and
outside of it. For example one of the commons-io vulnerabilities is
brought in via
[performance-analyzer-rca](https://github.com/opensearch-project/performance-analyzer-rca)
however performance-analyzer-rca is built as part of the
[performance-analyzer
](https://github.com/opensearch-project/performance-analyzer) (Different
repo) build process and requires injecting a patch into the build.gradle
that references another patch.