Fix /usr/src permissions (#30998)
Once all of these teething issues are fixed, this should probably
become a reusable pipeline, as it is tricky.

Fix permissions on the /usr/src/${{package.name}} directory, as mktemp
creates it as 700, despite contents being 644 or 755 as
appropriate. Without this fixup gdb does not work with source code
access for non-root users. Discovered in testing a custom image, as
non-root, with openssl-dbg pre-installed.

Fixes:
- https://github.com/wolfi-dev/os/pull/30691