Prevent rebuild of libxcrypt breaking the world
Currently if one rebuilds libxcrypt, circular dependency would be
introduced by melange sca:

```
diff libxcrypt-4.4.36-r7.apk libxcrypt.yaml
--- libxcrypt-4.4.36-r7.apk
+++ libxcrypt.yaml
@@ -8,5 +8,7 @@
 commit = 47793bfd95deb14489d0a0d187ff9abbd31f970a
 builddate = 1716471930
 license = GPL-2.0-or-later AND LGPL-2.1-or-later
+depend = so:libc.so.6
+depend = so:libcrypt.so.1
 replaces = libcrypt1<2.38-r15
 datahash = feb225bcfafd8f54c209f0aea74ecd9f55c652cdc756801b2c51ddd76cf2cc82
```

Disable depends generation to prevent this.

This is done because whilst libxcrypt is the one that ships
libcrypt.so.1, such provides with depends on libxcrypt is declared in
glibc libcrypt1 package. Due to same-origin preference in package
resolution in apko, it is impossible to currently move the provides to
libxcrypt package. As it results in hodling back upgrades of glibc.