Add pipeline for populating /etc/passwd,group; use in py3-breezy test
Convert the blob of code that adds a user to /etc/passwd into
a pipeline.  The pipeline allows you to ensure that the user is
present in /etc/passwd and /etc/group when you execute the content
provided.

The test here will now pass like:

    make test/py3-breezy MELANGE_EXTRA_OPTS=--runner=bubblewrap

It feels like py3-breezy might not be the only package who have
hit this.  I recently saw it in libreoffice also.
I'm not updating libreoffice here to skip 1 hour builds in any
discussion that  might occur. That can be done in follow-on.