py3-seaborn - multiver. py3-python-dateutil include dateutil-zoneinfo.tar.gz
py3-seaborn was previously a 'pip install' package which would
conflict with anything else and included pandas inside it and other
things downloaded from the internet.

Now it has dependencies and we build it.

py3-python-dateutil is updated here because 'import seaborn' would
previously throw an ugly warning:

    usr/lib/python3.12/site-packages/dateutil/zoneinfo/__init__.py:26:
       UserWarning: I/O error(2): No such file or directory
         warnings.warn("I/O error({0}): {1}".format(e.errno, e.strerror))

Now we are building the "zoneinfo" tarball and verifying that it is
there.