ruby3.2-kubeclient - Loosen restriction on recursive-open-struct version
ruby3.2-kubeclient is currently not passing a test.
That means it can't get any new version should one be created.
I hit this when trying to create ruby 3.3 versions of it.

If this does not work, then we will need to create a
ruby3.2-recursive-open-struct-1.3 package and depend on it.

I proposed upstream https://github.com/ManageIQ/kubeclient/pull/640.

Fixes: https://github.com/wolfi-dev/os/issues/31938