php modules - Drop virtual provides , replace remaining 8.X occurences.
We currently have 3 packages of each php module, one for each version of
PHP.  For amqp, that is:

   php-8.1-amqp
   php-8.2-amqp
   php-8.3-amqp

Each of these has a dependency on the corresponding php-8.X.

Before this commit, the module for the newest php version was also
providing 'php-amqp', such that if something just wanted 'php-amqp',
it would get php-8.3-amqp.

That is somewhat fragile though and would require us to remove the
provides from an 8.3 package when 8.4 is released.
(see https://github.com/wolfi-dev/os/pull/16420)

We don't want to have to re-build php-8.3-* modules in order to add
php-8.4 versions.

The other option here is to use provider-priority as we do in python.
But in order to get there, we would probably have to remove all the
previously published apks as they would 'provide' an explicit version.