Bumping selenium server to use jdk-17 (#32791)
<!---
Provide a short summary in the Title above. Examples of good PR titles:
* "ruby-3.1: new package"
* "haproxy: fix CVE-2014-123456"
-->
The selenium server in
https://github.com/SeleniumHQ/selenium/blob/cde43bcc398813a36142f01e026a34e0d71d1cbd/MODULE.bazel#L151
uses jdk-17. This PR is to bump the current version of jdk from 11 to
17. This allows the building of docker-selenium images
<!--
Please include references to any related issues or delete this section
otherwise.
 -->

Fixes:

Related:

### Pre-review Checklist

<!--
This checklist is mostly useful as a reminder of small things that can
easily be
forgotten – it is meant as a helpful tool rather than hoops to jump
through.

At the moment of this PR you have the most information on what all the
change
will affect, so please take the time to jot it down.

Put an `x` in all the items that apply, make notes next to any that
haven't been
addressed, and remove any items that are not relevant to this PR.

-->