dotnet-8: use offline build
Dotnet-8 tag 8.0.10 or older cannot be built anymore,
as Microsoft decided to retire the following repositories:

  * https://pkgs.dev.azure.com/ms/BuildXL/_packaging/BuildXL.Selfhost/nuget/v3/index.json
  * https://pkgs.dev.azure.com/ms/BuildXL/_packaging/BuildXL/nuget/v3/index.json

The repositories were replaced with:

  * https://pkgs.dev.azure.com/mseng/PipelineTools/_packaging/BuildXL.External.Dependencies/nuget/v3/index.json

The problem is that the tagged releases of dotnet <= 8.0.10 will never be able
to build anymore, as tags cannot be replaced.

The fix has landed to https://github.com/dotnet/msbuild/pull/10838, and the
following tagged releases should be buildable again.

See: https://github.com/microsoft/BuildXL/issues/1343

But, if we leverage the offline build, after the prep.sh has been run,
there is no required connection to the inexistent repositories, thus the build will
work for the old tags too.