ruby-3.3: Fix ruby 3.3.6 build and related improvements
Right now ruby-3.3 of version (3.3.5) is not building all of the gems it
installs as it fails silently due to missing deps.  In ruby 3.3.6 bundles
with C extensions that don't build will not be installed.  This changes fixes
things so that bundled gems are installed. But also trims down the bundled gems
that are installed to avoid conflicts with gems that we already had packages
for. For example ruby-3.3 (3.3.5) bundles rexml version 3.3.6 and our
ruby3.3-rexml modules packages version 3.3.9.  ruby-3.3.6 however bundles rexml
version 3.3.9 and causes a conflict between ruby-3.3 and ruby3.3-rexml.

Unfortunately we cannot include the rexml and net-imap packages as
dependecies for ruby to copmlete the same set of packages as it causes a
loop when determining build order.

In the future ruby packages should not bundle any of it's gems and
should be bootstrapped with all gems from source.