Adjust amazon-cloudwatch-agent and -operator to use tags.
amazon-cloudwatch-agent and amazon-cloudwatch-agent-operator
both are creating tags without creating releases in github.

 https://github.com/aws/amazon-cloudwatch-agent/tags
 https://github.com/aws/amazon-cloudwatch-agent-operator/tags

The agent is making releases more consistently from tags.
The operator has not made a release newer than v1.6.0 (Jul 30, 2024)
but has tags for 1.7.0, 1.8.0, 1.9.0 (yesterday).

This was noticed when the --auto-instrumentation-nodejs-image flag
was found to be missing from the manager binary (built by -operator).