Don't test for .pc files in /usr/local (#33608)
After some discussion with @xnox I realize that we don't actually want
to be testing for .pc files in /usr/local as that is a way for the user
to override the operating system's .pc files. Subsequently, we should
drop the change to check for them in /usr/local and also remote the
tests that will fail because they use /usr/local and consider those a
bug in the package.