fix(py3-conda-libmamba-solver): Cherrypick upstream commits to add support for libmamba 2.0
We have already released packages updates to libmamba and subpackages which is a major version jump from 1.5 to 2.0.

This caused failures in the conda packages and images due to py3-conda-libmamba-solver not yet supporting libmamba 2.0.

There is not yet an upstream release of conda-libmamba-solver with mamba 2.0 support for there is a release candidate
branch @ https://github.com/conda/conda-libmamba-solver/tree/24.11.0rc awaiting release.

We have broken dependent packages so we cannot wait for upstream to release.

As such I have cherrypicked the relevant 2.0 support commits from the 24.11.0rc branch.

This resolves the issues seen with conda and libmamba.

Full diff for conda-libmamba-solver @  https://github.com/conda/conda-libmamba-solver/compare/24.9.0...24.11.0rc

https://github.com/mamba-org/mamba/issues/3506 was the most helpful in finding the work to support libmamba 2.0
Signed-off-by: philroche <phil.roche@chainguard.dev>