abseil-cpp: Add GCC spec file to workaround GCC bug
abseil-cpp and packages that use it began to FTBFS after the introduction of
openssf-compiler-options, specifically its addition of the
-fno-delete-null-pointer-checks flag. This is believed to be a GCC bug:
  https://gcc.gnu.org/bugzilla/show_bug.cgi?id=71962

Unfortunately it isn't possible to "undo" this flag by appending the
inverse flag (-fdelete-null-pointer-checks). So here we introduce a
new spec file for abseil-cpp builds that will filter out the
problematic flag from the openssf spec file. Fixes #34075.

This commit uses the spec file to build abseil-cpp itself, and
ships it in the abseil-cpp-dev apk so abseil-cpp's reverse
dependencies can be updated to use it.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>