Drop virtual provides from php module packages
Some of the php modules had virtual provides. The intent of such
a thing would be so that a process could 'apk add php-amqp' and
get a sane result.

The implementation had a problem where by doing so would choose
whichever php-8.X-amqp version had the newest package version.
That is almost certainly _not_ what the process wanted (one day to get
php-8.2-ampq and the next to get php-8.1-ampq).

https://github.com/wolfi-dev/os/issues/32874 describes the problem
and a solution. https://github.com/wolfi-dev/os/pull/34680 implemented
that solution.  In the PR we realized that the provides were so
inconsistent that it would be had to use the. So then, given the
option of making them consistent or dropping them, we chose to drop
them.