fix(apache-nifi.yaml): Catch error when testing apache-nifi server start script
More accurately represent a correct apache-nifi server start - just as our container images do.

The start script will not exit non zero even if it fails. Previous test would always pass as long
as the script existed and was executable.

As such we need to ensure we start the same we to with the apache-nifi container image entrypoint
by setting the expected environment variables and by ensuring the expected log message is present.

Signed-off-by: philroche <phil.roche@chainguard.dev>