Withdraw lua-rrd-1.8.0-r0 due to it providing liblua-5.4.so
The 1.8.0-r0 version of lua-rrd provided so:liblua-5.4.so
That causes install problems for anything with a dependency
so:liblua-5.4.so as it might get the lua-rrd version.

The 1.9 versions are still providing it, but it gets marked
as 'vendored' now so it doesn't get selected.

It still should not provide the paths below, as it will then
conflict with lua5.4 or that has a liblua-5.4 dependency.

   # apk info -L lua-rrd
   lua-rrd-1.9.0-r1 contains:
   usr/lib/lua5.4/liblua-5.4.so.0
   usr/lib/lua5.4/liblua-5.4.so.0.0.0
   usr/lib/lua5.4/liblua.a
   usr/lib/lua5.4/liblua.la
   usr/lib/lua5.4/liblua.so