Remove AWX and Ansible operator packages (#34966)
This PR removes:
* awx-operator, whch depends on
* ansible-operator

because the Ansible operator can only run with non-latest 2.31.0 version
of the `requests` Python package, ending up needing an unguarded
package. And because it's a runtime dependency of the AWX operator, that
applies to that too.

Details available
[here](https://github.com/operator-framework/ansible-operator-plugins/pull/67#issuecomment-2189164688).
Code path non supported by later versions of requests is
[here](https://github.com/ansible/ansible-runner-http/blob/631cb47044d792e9e16ae924d26711d1346e3712/ansible_runner_http/events.py#L10).

---------

Signed-off-by: Massimiliano Giovagnoli <me@maxgio.it>