withdraw packages built by rrdtool version 1.8.0-r0
The most problematic package here is lua-rrd-1.8.0-r0
which copied a build of lua5.4 libraries into usr/lib/
and thus belives that it 'provides = so:liblua-5.4.so.0=0'.

Because the latest builds of lua5.4 do not actually have that
provide (they should), hilarity insues and things built against
liblua5.4 will install an out of date version of lua-rrd to
get it.

Stop the madness.