Add libquadmath as runtime dep for gcc (#35164)
This is silly, but the alternative to fix this will take a while, so I
am proposing this in the meantime. We currently (incorrectly) assume
that we will be able to solve for identical package sets across
architectures, but that's not correct for per-arch things like
libquadmath. I want to unwedge some things with this change, which I can
revert once we fix these assumptions downstream.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>