Fix(build): add patch to remove remove signing_key from gemspec
Our ruby/build pipeline delete all the lines contains 'signing_key' by default since it leads to build failures as the key is not available in the build environment and is not needed. ref: https://github.com/chainguard-dev/melange/blob/main/pkg/build/pipelines/ruby/build.yaml\#L42

But for version 2.0.3 and later it resutls in syntax error, unexpected end as it removes the condition line but not the full condition.

Signed-off-by: debasishbsws <debasishbsws.dev@gmail.com>