tensorflow-core: Drop python-version-specific headers from /usr/include
Currently we build for 1 python version and copy all of the headers
installed into the python-version-specific path into /usr/include. Once
we add python multiversion support, there will be a conflict. Which
python version's headers should we prefer?

Well, no wolfi packages depend on tensorflow-dev, and the extra headers
that we copy from the python subdir all seem to be for 3rd-party sources.
Let's just drop them for now and see if anyone screams.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>