libmamba: switch to py/pip-build-install pipeline
pip builds need to find the .cmake files that will have been staged under
${{targets.destdir}} by the main package pipeline, which requires setting
the libmamba_DIR environment variable. When we were building in a 'run:'
stage, we just set that environment variable on the command line.
Unfortunately, no such facility seems to exist for the py/pip-build-install
pipeline. ${{X.Y}} variables are not expanded in an 'environment:' block.
So here use a 'run:' block to create python interpreter wrappers to set
this variable before calling the real interpreter.

We also no longer set CMAKE_PREFIX_PATH because files get installed
in the correct places without it.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>