libmamba: use semver tags, split into separate yaml files, update to 2.0.5
We were tracking the date-based tags, but these tags are not
always forward-moving. Switch to the semver tags instead.

Now, there isn't a single stream of semver tags in this project, rather
there are tags per binary - e.g. libmamba-2.0.5, libmambapy-2.0.5 and
micromamba-2.0.5. These tags appear to always point to the same commit
hash for each release, but I'm not sure we can assume that will always
be the case. Let's just treat them as separate projects.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>