python3.10, python-3.11 - build tkinter with python
This is similar to the python 3.11 and python-3.12 change to
build tkinter inside python.

The big difference is that we have to use tcl and tk at major
version 8 because python3.10 and python-3.11 do not support
tcl/tk at version 9.

The proper way to do this would be to version tcl and tk
(tcl-8.yaml, tk-8.yaml).

In order to correctly use tkinter (and pass the provided test),
we have to add explicit runtime tcl and tk dependencies because melange
does not identify the dependency that