simplify lint.sh, don't rely on build order (#36324)
This also removes some logic that we put in a ~year ago to supported
nested package dirs, but then never actually took advantage of. If we
want that we can add it back.

This removes `make list` which was only used in linting, and replaces it
with `make check-bootstrap` which makes it clear it's only there to
determine a bootstrap order, and nothing else.

---------

Signed-off-by: Jason Hall <jason@chainguard.dev>