don't require build maintainers approval (#36325)
Historically, we imposed a limit on who could approve changes in
`Makefile` and `.github/workflows/` to prevent PRs from breaking CI.

Now that builds happen in elastic-build, it's a lot harder for randos to
break the build (not impossible; _this does not constitute a
challenge_), and we should let others contribute to the Makefile and GHA
workflows. The blast radius of changes to those are a lot smaller now.

Signed-off-by: Jason Hall <jason@chainguard.dev>