ruby 3.4: interpreter & gem batch 1 (#36408)
This was created by renaming the 3.3 file to 3.4.

Turning the gemspec patches into files. See
https://github.com/wolfi-dev/os/blob/main/ruby3.2-bouncy-castle-java/0001-add-gemspec.patch


Providing the base64, csv, and drb package from the main ruby package
because they now conflict and some packages have implicit dependencies
on them. Which means we either need to spot update failing packages or
provide it from the interpreter that is already installing these gems.