rsyslog: Use relative symlink
This works around what seems to be a bug in apko when installing a
directory in a parent directory that happens to be a symlink. The
relative syntax appears to work but this absolute syntax doesn't.

Rather than trying to fix this in apko (which, I'm not sure if it's a
bug or not, because the error is really coming from os stdlib), let's
just make this relative.

Other distros seem to do this by default, so we may want to do this in
wolfi-baselayout instead, but I'm not confident enough in my distro
skills to propose that.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>