remove py3-pywinpty which should never have been added, as it require… (#36706)
We had a failing package update PR that popped up in sustaining:
 - https://github.com/wolfi-dev/os/pull/31028

On closer inspection, this was a win32 package, and should never have
been added to wolfi. Whilst it may have been compiling before, it was
non-functional. It's highly likely it was wrongly imported when
jupyter-server was built / imported.

**Please see the following GH issue comments for more information:**
- https://github.com/wolfi-dev/os/pull/31028#issuecomment-2527809868

This PR cleans up the package, as well as adds it to the withdrawn list.
There are no other packages (in wolfi or elsewhere), or images with
dependencies on this broken package.

Signed-off-by: Mark McCormick <mark.mccormick@chainguard.dev>