Add python-as-wrapper package.
Many programs or build systems really want to use 'python3',
but we really want them to use 'python3.X'.

This wrapper package provides a compromise that gets rid of the
"sticky" results of calling 'python3'.  The big change is that
now a program that uses 'python3' to write its scripts will
write a shbang as '#!/usr/bin/python3.XX'

See python-as-wrapper/README.md for more information.