nodetaint/GHSA-27wf-5967-98gx fix (#36568)
This is a good example as to how cluttered a package can get when
attempting to remediate k8s dependencies with go mod / go/bump. It is a
simple version bump but the dependencies are so tightly coupled and with
no recursive dependency updating there can be trail and error in finding
everything needing to be updated. Anyway, version bumped, epoch bumped.