airflow: pin to python3.12, due to FTBFS with python3.13 (#37113)
Currently airflow is failing to build from source in post-submit, due to
python3 changing from 3.12 to 3.13.

This change happened after successful pre-submit and prior to successful
build in post-submit. This race is possible as we do not require "up to
date" PR prior to merging, which would be impractical to enforce.

In the future, prior to flipping default, we should detect FTBFS and pin
packages to 3.12 prior to upgrading the default.