feat: Add kubernetes daemonsets for fluentd 1.17 (#36618)
Adds Kubernetes DaemonSets for fluentd. Note that fluentd is vendored in
the bundle so is not included as a package dependency. Upstream includes
two separate installations of fluentd which is not necessary. Add for
all supported Ruby versions (3.1-3.3) and 3.4 (in preview)

There isn't value in splitting away the entrypoint scripts or
configuration for each so those have been bundled together

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>