fix(druid-pombump): Set netty4.version property instead of netty-common version directly
After analyzing the druid pom structure, and the stacktrace linked in the stale druid
issue, realized that we were getting a class compatibility issue with netty-common,
specifically the Timer class. This is because every other netty dependency was a
different version that netty-common. By setting the netty4.version property, we guarantee
that all netty libs are on the same version and are therefore compatible. And the reason
we added the pombump dependency in the first place (probably to solve a CVE) is preserved.

Signed-off-by: Samuel Dacanay <sam.dacanay@chainguard.dev>