Fix Druid pombump configuration (#37343)
After analyzing the druid pom structure, and the stacktrace linked in
the stale druid
issue, realized that we were getting a class compatibility issue with
netty-common,
specifically the Timer class. This is because every other netty
dependency was a
different version that netty-common. By setting the netty4.version
property, we guarantee
that all netty libs are on the same version and are therefore
compatible. And the reason
we added the pombump dependency in the first place (probably to solve a
CVE) is preserved.

The class compatiblity issue was causing the imagetest for druid to
fail, because services could not correctly come up with that issue,
which is part of the Druid startup process.

Fixes https://github.com/chainguard-dev/image-release-stats/issues/3075

Signed-off-by: Samuel Dacanay <sam.dacanay@chainguard.dev>