fix(aws-ebs-csi-driver): Add missing dependencies, create needed symlink (#37471)
The AWS EBS CSI driver depends on various utilities that weren't
provided. Add them as runtime dependencies

Also symlink the driver to /bin so that it's found (this is where
upstream places it and where the chart expects it). This package isn't
useful without this link so don't relegate to a compat package

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>