nodejs - be explicit about which python is used in build environment. (#37223)
When 'python3' and 'py3-jinja2' are installed in the build environment,
there is not currently a way to guarantee that the 'python3' (a virtual
package provided by python3.12, python3.13...) will be the same python
version that fulfills the virtual package 'py3-jinja2'.

The easiest way to guarantee this is to pin the python version as done
here.

python-3.12 (eol 2028) will outlive any of the currently available
nodejs versions. So it would be sufficient for the 'py-version' value in
all of these changes. Moving the last node lts (22) to 3.13 will give us
a neweer python moving forward.