gobject-introspection - use specific python version, test g-ir-scanner
gobject-introspection build writes programs in /usr/bin/
g-ir-scanner and g-ir-annotation-tool that used '#!/usr/bin/env python3'

It installs python files and modules into /usr/lib/gobject-introspection
and successfully finds that at runtime.  However, it built a c extension
_giscanner.cpython-312-x86_64-linux-gnu.so which will only load with
a specific python (the one that it was built against).

So, update the shbang on g-ir-scanner and g-ir-annotation-tool to
reference the specific python it was built against.