audit - update prefix for install to /usr rather than /usr/local (#37646)
The change here simplifies flags setting in configure and also sets the
--prefix to /usr.

Setting to /usr means that audit's python package is installed into
/usr/lib/python3.XX/site-packages where it is then picked up as a
melange SCA depends.

This _does_ mean that the python that is built rolls with forward with
whatever 'python3-dev' happens to install. For now that seems fine.