pylint - specify a python version (#37291)
When building a python package, it will install into a specific
usr/lib/python3.XX/site-packages. Thus, there is a runtime dependency
on that specific python version.   Melange identifies and records
that correctly.

However, for python runtime dependencies melange does not record any
dependencies. The solution in the past was to record those dependencies
as 'py3-<pkgname>'. The problem with that solution is that py3-pkgname
is a virtual package that will move versions over time (ie, from 3.12 to
3.13).

The result was that pylint had a python3.12 dependency, but py-astroid
(and others) would install 3.13. We'd have 2 installed pythons and no
working pylint.