apache-pulsar: adding missed runtime dependencies / upgrade bouncycastle (#38057)
<!---
Provide a short summary in the Title above. Examples of good PR titles:
* "ruby-3.1: new package"
* "haproxy: fix CVE-2014-123456"
-->

<!--
Please include references to any related issues or delete this section
otherwise.
 -->

Fixes:

- Adds `openssl` and `procps` as runtime dependencies
- Builds with `openjdk-21` - external counterpart image uses 21
- Upgrades bouncycastle-bcfips to 2.x

Related: https://github.com/chainguard-dev/image-requests/issues/1654

### Pre-review Checklist

<!--
This checklist is mostly useful as a reminder of small things that can
easily be
forgotten – it is meant as a helpful tool rather than hoops to jump
through.

At the moment of this PR you have the most information on what all the
change
will affect, so please take the time to jot it down.

Put an `x` in all the items that apply, make notes next to any that
haven't been
addressed, and remove any items that are not relevant to this PR.

-->

#### For PRs that add patches
<!-- remove if unrelated -->
- [x] Patch source is documented

---------

Signed-off-by: Michael Paul <31833332+MickeyPvX@users.noreply.github.com>