conda: avoid py3.x-conda cross-contamination
Anything in the top-level directory of this package will end up getting
installed into each version of the python module. We already use the
`prevent-inclusion` parameter to prevent those files from being copied
in. Now that we're using "cleanup" as a holding area, we need to include
that as well.

This was causing the usr/bin/* files from all previously built modules
to be included in subsequent ones:

  # apk info -L py3.12-conda | grep cleanup
  WARNING: opening /work/packages: No such file or directory
  usr/lib/python3.12/site-packages/cleanup/3.10/bin/conda
  usr/lib/python3.12/site-packages/cleanup/3.11/bin/conda
  usr/lib/python3.12/site-packages/melange-out/py3.11-conda/usr/lib/python3.11/site-packages/cleanup/3.10/bin/conda

And, because melange adds dependencies for paths in the shebangs,
we were getting dependencies on additional python versions.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>